+++ Vocable/Common/CarouselGridLayout.swift
@@ -96,14 +96,6 @@ final class CarouselGridLayout: UICollectionViewLayout {
 
         static let zero: Self = .init(interRowSpacing: 0, interColumnSpacing: 0)
         static func uniform(_ spacing: CGFloat) -> Self { .init(interRowSpacing: spacing, interColumnSpacing: spacing) }
-
-        static func > (lhs: InterItemSpacing, rhs: InterItemSpacing) -> Bool {
-            return lhs.interRowSpacing + lhs.interColumnSpacing > rhs.interRowSpacing + rhs.interColumnSpacing
-        }
-
-        static func < (lhs: InterItemSpacing, rhs: InterItemSpacing) -> Bool {
-            return lhs.interRowSpacing + lhs.interColumnSpacing < rhs.interRowSpacing + rhs.interColumnSpacing
-        }
     }
 
     var interItemSpacing: InterItemSpacing = .zero {
@@ -375,7 +367,7 @@ final class CarouselGridLayout: UICollectionViewLayout {
 
     override func targetContentOffset(forProposedContentOffset proposedContentOffset: CGPoint, withScrollingVelocity velocity: CGPoint) -> CGPoint {
         guard let collectionView = collectionView, pagesPerSection > 0,
-              pageContentSize.width > 0 || interItemSpacing > .zero else { return proposedContentOffset }
+              pageContentSize.width > 0 else { return proposedContentOffset }
         let proposedCenterX: CGFloat
         if velocity.x > 1.0 {
             proposedCenterX = collectionView.bounds.maxX
